Fix tests and move StatsDateFormatter into Utils package
